+++ src/util.c
@@ -341,6 +341,18 @@ char *util_get_config_dir(void)
 }
 
 /**
+ * Retrieves the data directory path according to current used profile.
+ * Returned string must be freed.
+ */
+char *util_get_data_dir(void)
+{
+    char *path = g_build_filename(g_get_user_data_dir(), PROJECT, vb.profile, NULL);
+    create_dir_if_not_exists(path);
+
+    return path;
+}
+
+/**
  * Retrieves the length bytes from given file.
  *
  * The memory of returned string have to be freed with g_free().